requirements.txt
Django==3.2.12
django-extensions==3.1.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
kombu==5.2.1
Pillow==7.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
virtualenv==20.0.17
gunicorn
mock

